opened assets,components and pages folders, moved the reusable components into components, the general things to assets. Also changed the imports accordingly
